refactor(td-utils): convert td-utils to TypeScript project
