+++ .gitignore
@@ -7,6 +7,7 @@ node_modules
 # dist directory should be included only in the Package but not Repository
 dist
 out
+build
 
 # logging
 logs.json
