+++ node/td-utils/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    presets: [["@babel/preset-env", { targets: { node: "current" } }], "@babel/preset-typescript"],
+};
