+++ node/td-utils/index.ts
@@ -1,0 +1,1 @@
+export { detectProtocolSchemes } from "./src/detectProtocolSchemes";
