+++ node/td-utils/package.json
@@ -2,12 +2,14 @@
     "name": "@thingweb/td-utils",
     "version": "1.0.0",
     "description": "The package provides utility tools for TDs.",
-    "main": "index.js",
+    "main": "./build/index.js",
+    "types": "./build/index.d.js",
     "scripts": {
+        "build": "tsc -b",
         "test": "jest",
         "test:coverage": "jest --coverage",
-        "bundle": "npx browserify ./index.js -s tdUtils -o ./dist/web-bundle.js && npx terser ./dist/web-bundle.js -o ./dist/web-bundle.min.js",
-        "prepare": "npm run bundle"
+        "bundle": "npx browserify ./build/index.js -s tdUtils -o ./dist/web-bundle.js && npx terser ./dist/web-bundle.js -o ./dist/web-bundle.min.js",
+        "prepare": "npm run build && npm run bundle"
     },
     "repository": {
         "type": "git",
@@ -15,9 +17,16 @@
         "directory": "node/td-utils"
     },
     "devDependencies": {
+        "@babel/core": "^7.24.7",
+        "@babel/preset-env": "^7.24.7",
+        "@babel/preset-typescript": "^7.24.7",
+        "@types/jest": "^29.5.12",
+        "@types/lodash": "^4.17.6",
+        "babel-jest": "^29.7.0",
         "browserify": "^17.0.0",
         "jest": "^28.1.3",
-        "terser": "^5.14.2"
+        "terser": "^5.14.2",
+        "wot-thing-description-types": "^1.1.0-09-November-2023"
     },
     "author": "Eclipse Thingweb <thingweb-dev@eclipse.org> (https://thingweb.io/)",
     "license": "EPL-2.0 OR W3C-20150513",
